+++ scrapers/match_odds_scraper.py
@@ -152,7 +152,10 @@ def save_year_match_odds_to_csv(
     year: str, match_odds: dict[str, dict[str, float]]
 ) -> None:
     """Write one CSV row per match: match_url, team_a, odd_a, team_b, odd_b."""
-    out_file = Path("thunderbird_match_odds") / f"{year}_thunderbird_match_odds.csv"
+    out_file = (
+        Path("scraped_data/thunderbird_match_odds")
+        / f"{year}_thunderbird_match_odds.csv"
+    )
     out_file.parent.mkdir(parents=True, exist_ok=True)
 
     with out_file.open("w", newline="", encoding="utf-8") as f:
