+++ valorant_match_predictor/main.py
@@ -4,10 +4,6 @@ import numpy as np
 from sklearn.preprocessing import StandardScaler
 
 from typing import Callable
-from collections import defaultdict, Counter
-from pathlib import Path
-
-import csv
 
 from valorant_match_predictor import (
     DATAFRAME_BY_YEAR_TYPE,
@@ -284,6 +280,7 @@ def get_match_predictor_model(
 
 
 def compute_odds(yearly_probabilities):
+    odds = []
     for probabilities in yearly_probabilities:
         eps = 1e-6
         probs = np.clip(probabilities.squeeze(1).cpu().numpy(), eps, 1 - eps)
@@ -299,7 +296,9 @@ def compute_odds(yearly_probabilities):
         )
 
         pd.set_option("display.float_format", "{:.2f}".format)
-        print(df.head())
+        odds.append(df)
+
+    return odds
 
 
 def train(
@@ -324,38 +323,30 @@ def test(
     # transformed_data = transform_data(dataframes_by_year)
 
     pred_match_probs = []
-    print(thunderbird_match_odds[:10])
 
-    # players_stats = year_data["players_stats"]["team_players_stats"]
-    # matchups_data = year_data["matches"]["teams_matchups_stats"]
+    for match_url, odds in thunderbird_match_odds.items():
+        print(match_url)
+        print(odds)
+        print()
+        # get the recent player stats before this game (likely through web scraping)
+        # get recent matchup stats
 
-    # team_a_tensor, team_b_tensor, win_probabilities, _ = create_match_input_tensors(
-    #     pr_model, players_stats, matchups_data
-    # )
+        # players_stats = year_data["players_stats"]["team_players_stats"]
+        # matchups_data = year_data["matches"]["teams_matchups_stats"]
 
-    # team_a_mask = ~(torch.isnan(team_a_tensor).any(dim=1))
-    # team_b_mask = ~(torch.isnan(team_b_tensor).any(dim=1))
-    # team_a_tensor = team_a_tensor[team_a_mask]
-    # team_b_tensor = team_b_tensor[team_b_mask]
+        # team_a_tensor, team_b_tensor, win_probabilities, _ = create_match_input_tensors(
+        #     pr_model, players_stats, matchups_data
+        # )
 
-    # probabilities = match_model(team_a_tensor, team_b_tensor)
-    # pred_match_probs.append(probabilities)
+        # team_a_mask = ~(torch.isnan(team_a_tensor).any(dim=1))
+        # team_b_mask = ~(torch.isnan(team_b_tensor).any(dim=1))
+        # team_a_tensor = team_a_tensor[team_a_mask]
+        # team_b_tensor = team_b_tensor[team_b_mask]
 
-    # predicted_odds = compute_odds(pred_match_probs)
+        # probabilities = match_model(team_a_tensor, team_b_tensor)
+        # pred_match_probs.append(probabilities)
 
-
-def load_year_match_odds_from_csv(year: str) -> dict[str, dict[str, float]]:
-    in_file = Path("thunderbird_match_odds") / f"{year}_thunderbird_match_odds.csv"
-    match_odds = defaultdict(dict)
-    with in_file.open(newline="", encoding="utf-8") as f:
-        reader = csv.DictReader(f)
-        for row in reader:
-            match_url = row["match_url"]
-            team_a, team_b = row["team_a"], row["team_b"]
-            odd_a, odd_b = float(row["odd_a"]), float(row["odd_b"])
-            match_odds[match_url] = {team_a: odd_a, team_b: odd_b}
-
-    return match_odds
+        # predicted_odds = compute_odds(pred_match_probs)
 
 
 if __name__ == "__main__":
@@ -368,4 +359,5 @@ if __name__ == "__main__":
     set_pandas_options()
     pr_model, match_model = train(years=["2022", "2023"])
     thunderbird_match_odds = load_year_match_odds_from_csv("2024")
+    # print(thunderbird_match_odds)
     test(pr_model, match_model, thunderbird_match_odds)
